requirements.txt
websockets
pytest>=9.0.2
pyserial>=3.5,<4.0
